+ ajout article

+++ edit_script.php
@@ -1,9 +1,9 @@
 <?php
 session_start();
 include('db_class.php');
-if (isset($_POST['mod']) AND isset($_POST['article-name']) AND isset($_POST['article-desc']) AND isset($_POST['article-price']) AND isset($_POST['stock-name']) AND $_SESSION['droits'] AND isset($_GET['id']) AND isset($_POST['article-cat'])) {
+if (isset($_POST['mod']) AND isset($_POST['article-name']) AND isset($_POST['article-desc']) AND isset($_POST['article-price']) AND isset($_POST['stock-name']) AND $_SESSION['droits'] AND isset($_POST['article-cat'])) {
     $db = new MyDB();
-    if ($_POST['mod'] === '1') {
+    if ($_POST['mod'] === '1' AND isset($_GET['id'])) {
         if ($_FILES['article-image']['tmp_name'] AND getimagesize($_FILES['article-image']['tmp_name'])) {
             $mime = mime_content_type($_FILES['article-image']['tmp_name']);
             $b64img = base64_encode(file_get_contents($_FILES['article-image']['tmp_name']));
@@ -25,9 +25,24 @@ if (isset($_POST['mod']) AND isset($_POST['article-name']) AND isset($_POST['art
             $req->bindValue(':idCat', $_POST['article-cat']);
             $req->bindValue(':id', $_GET['id']);
         }
-    } else {
+    } elseif ($_POST['mod'] === '2' AND isset($_GET['id'])) {
         $req = $db->prepare("DELETE FROM Articles WHERE idProd=:id;");
         $req->bindValue(':id', $_GET['id']);
+    } elseif ($_POST['mod'] === '3' AND $_FILES['article-image']['tmp_name'] AND getimagesize($_FILES['article-image']['tmp_name'])) {
+            $mime = mime_content_type($_FILES['article-image']['tmp_name']);
+            $b64img = base64_encode(file_get_contents($_FILES['article-image']['tmp_name']));
+            $req = $db->prepare("INSERT INTO Articles (nom, quantite, prix, description, b64img, idCat, mime) VALUES (:nom, :quantite, :prix, :description, :b64img, :idCat, :mime);");
+            $req->bindValue(':nom', $_POST['article-name']);
+            $req->bindValue(':quantite', $_POST['stock-name']);
+            $req->bindValue(':prix', $_POST['article-price']);
+            $req->bindValue(':description', $_POST['article-desc']);
+            $req->bindValue(':b64img', $b64img);
+            $req->bindValue(':idCat', $_POST['article-cat']);
+            $req->bindValue(':mime', $mime);
+            $req->bindValue(':id', $_GET['id']);
+    } else {
+        $_SESSION['error'] = 'Erreur dans le formulaire.';
+        header('location: admin.php');
     }
     $req->execute();
     $_SESSION['error'] = $db->lastErrorMsg();
